requirements.txt
flappy_bird_gymnasium
tensorflow
pytorch
torch
torchvision 
torchaudio --index-url https://download.pytorch.org/whl/cu124
pyyaml